+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#148](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/148))
 - weapon "Staff" having a short range of 2 instead of 4
   ([#149](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/149))
+- the chat log not scrolling to the bottom on weapon attacks
+  ([#150](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/150))
 
 ## [0.11.2] - 2022-01-22